Moved telemetry into a contribution
